+++ src/js/class/Dom.js
