+++ src/js/class/Filter.js
