+++ src/js/class/Helper.js
